+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## 0.2.0
 - Template escaping is now enabled by default in `$html`
+- Removed `escapeQuotes` method, using traditional escaping is safe for quotes
 - `$html` now returns an object with `#toString()` instead of a string
 
 ## 0.1.0